Bug fix for deleted guardian
Could still edit child information of a deleted guardian. Didnt think we
should be able to do that. Let me know if you have an issue with that
fix.

+++ ChildcareApplication/ChildcareApplication/AdminTools/Link_DeLinkChild.xaml.cs
@@ -58,7 +58,8 @@ namespace AdminTools {
                             WPFMessageBox.Show("Guardian with ID: " +pID + " does not exist.");
                     } else if (linked == 1) {//delink child
                         bool connExists = conDB.ConnectionExists(pID, childID);
-                        if(connExists)
+                        bool guardianExists = gdb.GuardianNotDeletedAndExists(pID);
+                        if(connExists && guardianExists)
                             conDB.DeleteAllowedConnection(childID, pID);
 
                         else
